requirements.txt
streamlit
flask
werkzeug
pycparser
python-docx
pdfminer.six
pytest
pytest-html
pytest-xdist
junit-xml
lxml
openai